Refactoring: Split lib/featureDiff.js
It's one component per file now in one folder.
