+++ lib/map.js
@@ -4,7 +4,7 @@ import { featureCollection } from '@turf/helpers';
 
 import { getBoundingBox, getBounds, getFeatureBBOX } from './helpers';
 import { cmap } from './render';
-import { displayDiff } from './featureDiff';
+import { renderFeatureDiff } from './featureDiff/renderFeatureDiff';
 
 const LAYERS_KEY = {
   'added-line': { added: true, ways: true },
@@ -1248,13 +1248,14 @@ export class Map {
       }
     });
   }
+
   selectFeature(feature) {
     var featureMap = this.result.featureMap;
     var featureId = feature.properties.id;
     var osmType = feature.properties.type;
 
     this.highlightFeature(featureId);
-    displayDiff(
+    renderFeatureDiff(
       featureId,
       featureMap,
       this.result.changeset.id,
@@ -1262,12 +1263,14 @@ export class Map {
       document.querySelector('.cmap-diff-tags'),
       document.querySelector('.cmap-diff-members')
     );
+
     document.querySelector('.cmap-diff').style.display = 'block';
     if (osmType === 'relation') {
       this.showRelationMembers(featureId);
     }
     cmap.emit('featureChange', osmType, featureId);
   }
+
   zoomToFeatures(features) {
     this.map.fitBounds(getFeatureBBOX(features), {
       linear: true,
@@ -1275,13 +1278,16 @@ export class Map {
       maxZoom: 18
     });
   }
+
   highlightFeature(featureId) {
     this.map.setFilter('highlight-line', ['==', 'id', featureId]);
     this.map.setFilter('highlight-point', ['==', 'id', featureId]);
   }
+
   selectMember(feature) {
     this.highlightRelationMemberFeature(feature);
   }
+
   highlightRelationMemberFeature(featureId) {
     this.map.setFilter('highlight-relation-member-line', [
       '==',
@@ -1294,11 +1300,13 @@ export class Map {
       featureId
     ]);
   }
+
   clearHighlight() {
     this.map.setFilter('highlight-line', ['==', 'id', '']);
     this.map.setFilter('highlight-point', ['==', 'id', '']);
     this.clearRelationMemberHighlight();
   }
+
   clearRelationMemberHighlight() {
     this.map.setFilter('highlight-relation-member-line', ['==', 'ref', '']);
     this.map.setFilter('highlight-relation-member-point', ['==', 'ref', '']);
@@ -1311,11 +1319,13 @@ export class Map {
         this.map.setLayoutProperty(layer, 'visibility', 'visible')
       );
   }
+
   clearFeature() {
     this.clearHighlight();
     clearDiff();
     cmap.emit('featureChange', null, null);
   }
+
   showRelationMembers(featureId) {
     Object.keys(this.relationMembersLayers).forEach(layerName => {
       let layer = this.relationMembersLayers[layerName];
